Finalized JS file with required updated.
Fixed the t-shirt section! I worked for multiple days on the issue of properly updating the designs based on the t-shirt selections. I discovered that I had to set the first option while hiding and showing the relevant designs. The answer was SUPER simple but not at all obvious. I did learn a ton while going through this process!